obj: handle ENOMEM during vector allocations
When missing, Handle ENOMEM during vector allocations
to avoid later crashing.

Ref: pmem/issues#5515
