+++ src/libpmemobj/tx.c
@@ -194,7 +194,8 @@ tx_action_add(struct tx *tx)
 	if (tx_action_reserve(tx, 1) != 0)
 		return NULL;
 
-	VEC_INC_BACK(&tx->actions);
+	if (VEC_INC_BACK(&tx->actions) == -1)
+		return NULL;
 
 	return &VEC_BACK(&tx->actions);
 }
@@ -710,6 +711,11 @@ tx_construct_user_buffer(struct tx *tx, void *addr, size_t size,
 		tx->redo_userbufs_capacity +=
 			userbuf.size - TX_INTENT_LOG_BUFFER_OVERHEAD;
 	} else {
+		uint64_t buffer_offset = OBJ_PTR_TO_OFF(ctx->p_ops->base,
+							userbuf.addr);
+
+		if (VEC_PUSH_BACK(&ctx->next, buffer_offset) != 0)
+			goto err;
 		operation_add_user_buffer(ctx, &userbuf);
 	}
 
@@ -1013,8 +1019,16 @@ pmemobj_tx_commit(void)
 		operation_start(tx->lane->external);
 
 		struct user_buffer_def *userbuf;
-		VEC_FOREACH_BY_PTR(userbuf, &tx->redo_userbufs)
-			operation_add_user_buffer(tx->lane->external, userbuf);
+		struct operation_context *ctx = tx->lane->external;
+		uint64_t buf_off;
+		VEC_FOREACH_BY_PTR(userbuf, &tx->redo_userbufs) {
+			buf_off = OBJ_PTR_TO_OFF(ctx->p_ops->base,
+						userbuf->addr);
+			if (VEC_PUSH_BACK(&ctx->next, buf_off) != 0)
+				FATAL("%s: failed to allocate the next vector",
+					__func__);
+			operation_add_user_buffer(ctx, userbuf);
+		}
 
 		palloc_publish(&pop->heap, VEC_ARR(&tx->actions),
 			VEC_SIZE(&tx->actions), tx->lane->external);
@@ -1921,7 +1935,11 @@ pmemobj_tx_xpublish(struct pobj_action *actv, size_t actvcnt, uint64_t flags)
 	}
 
 	for (size_t i = 0; i < actvcnt; ++i) {
-		VEC_PUSH_BACK(&tx->actions, actv[i]);
+		if (VEC_PUSH_BACK(&tx->actions, actv[i]) != 0) {
+			int ret = obj_tx_fail_err(ENOMEM, flags);
+			PMEMOBJ_API_END();
+			return ret;
+		}
 	}
 
 	PMEMOBJ_API_END();
